% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corr_df.R
\name{Corr_df}
\alias{Corr_df}
\title{Corr_df - Automated correction of PAM using mean for specifiedy classes.}
\usage{
Corr_df(df, pos1, pos2 = NULL)
}
\arguments{
\item{df}{data.frame - containing Control and PAM values with colnames "CTR" and "PAM"}

\item{pos1}{numeric - Column of data frame which should be used to correct by mean.}

\item{pos2}{numeric - Additional column of data frame which should be used to correct by mean.
Only necessary if the correction should use subgroups for the data in 'pos1'}
}
\value{
returns a data.frame with the corrected PAM values and prints the respective mean vlaues for each class or combination of classes.
}
\description{
visualizes the correction of PAM values by correction value.
}
\examples{
# load data
dat <- read.csv(system.file("extdata","exp_PAM.csv",package = "PAMcorrection"))
head(dat)

# correct df by means for "generation"
which(colnames(dat)=="generation")
corrected <-Corr_df(dat,pos1 = 3)

# use two classes (means for each generation depending on dilect)
which(colnames(dat)=="generation")
which(colnames(dat)=="dilect")
corrected <-Corr_df(dat,pos1 = 3,pos2 = 2)
}
\author{
Andreas Schönberg
}
