% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPAMcorr.R
\name{estPAMcorr}
\alias{estPAMcorr}
\title{estPAMcorr - Function for estimation of PAM correction.}
\usage{
estPAMcorr(ctr, pam, cf = NULL, al = -1, modef = T)
}
\arguments{
\item{ctr}{data.frame column - with Control values.}

\item{pam}{data.frame column - with PAM values.}

\item{cf}{numeric - desired correction factor.}

\item{al}{numeric - a vector of numeric values on the x axes to draw lines. Default=0}

\item{modef}{bolean - If TRUE uses input 'cf' to multiply PAM, if FALSE uses
substracts 'cf' from PAM. Default= TRUE}
}
\value{
returns a plot containing the original PAM values, control values and
corrected PAM values along the absolute difference between corrected PAM and
control values (green). Further returns mean and sum of this difference for
quality estimation of the result.
}
\description{
visualizes the correction of PAM values by correction value.
}
\details{
\itemize{
\item cf - it is recommended to calculate the mean of the diffenrence
between the original PAM and control values. Either use devision Control/PAM
for correction by factor and abs(PAM)-abs(Control) for correction by absolute values.
\item al - is used to visualise the borders of differnet classes for the values
}
}
\note{
For 'cf' it is recommended to calculate the mean of the diffenrence
between the original PAM and control values. Either use devision Control/PAM
for factor and abs(PAM)-abs(Control) for correction by absolute values.
}
\examples{
# load data
dat <- read.csv(system.file("extdata","exp_PAM.csv",package = "PAMcorrection"))
head(dat)

# visualize Control and PAM values
estPAMcorr(ctr = dat$CTR,pam = dat$PAM)

# view borders of dialect
estPAMcorr(ctr = dat$CTR,pam = dat$PAM,al = c(9,15,21,36))

# use a correction factor
cf <- mean(dat$CTR/dat$PAM)

estPAMcorr(ctr = dat$CTR,pam = dat$PAM,al = c(9,15,21,36),cf = cf)

# use absolute value for correction
cfa <- abs(mean(dat$CTR-dat$PAM))

estPAMcorr(ctr = dat$CTR,pam = dat$PAM,al = c(9,15,21,36),cf = cfa,modef = FALSE)
}
\author{
Andreas Schönberg
}
