% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMcorrection-Package.R
\docType{package}
\name{PAMcorrection - package}
\alias{PAMcorrection - package}
\title{PAMcorrection - a tool for PAM correction}
\description{
The \code{PAMcorrection} package is used to estimate and test correction values for the PAM using control values.
}
\details{
PAMcorrection
}
\author{
Andreas Schönberg
}
\keyword{package}
