% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPAMcorr.R
\name{plotPAMcorr}
\alias{plotPAMcorr}
\title{plotPAMcorr - Function for plotting of PAM correction.}
\usage{
plotPAMcorr(df, al = -1, yl = 2, sortby = NULL, titel = "PAM Correction")
}
\arguments{
\item{df}{data.frame - with column "CTR" for control values, "PAM" for pam values
and optional "PAM_corr" for corrected PAM values.}

\item{al}{numeric - a vector of numeric values on the x axes to draw lines. Default=0}

\item{yl}{numeric - ylim adjustment. Default=2}

\item{sortby}{character - column name to reorder the inoput datafram 'df'. IF set the argument 'al' will not be used.}

\item{titel}{character - desired maintitel for the plot . Default="PAMcorrection".}
}
\value{
returns a plot containing the original PAM (orange) and Control (blue) values along with the difference (grey).
If corrected PAM values are available those will be plotted in red along with the differnece between the corrected PAM and COntrol values.
Additionally prints the total Difference along with mean and standart divation for PAM or if given PAM corrected and Control
}
\description{
visualizes the correction of PAM values by correction value.
}
\note{
For visualization purposes the difference between the corrected PAM and Control is used as absolute values (negative values are plotted as positive).
Further the mean and sd is calculated for the absolute values.
}
\examples{
# load data
dat <- read.csv(system.file("extdata","exp_PAM.csv",package = "PAMcorrection"))
head(dat)

# plot control and PAM
plotPAMcorr(dat)

# adjust y limit
plotPAMcorr(dat,yl = 4)

# use vertical lines
abl <- seq(10,40,10)
plotPAMcorr(dat,yl = 4,al = abl)

# order values by PAM
plotPAMcorr(dat,sortby = "PAM")

# get correction value by mean
datf <-dat
datf$PAM_corr <-dat$PAM * mean(dat$CTR/dat$PAM) # factor
data <-dat
data$PAM_corr <-dat$PAM - mean(dat$PAM-dat$CTR) # factor

# plot correction and use titel
plotPAMcorr(datf,sortby = "PAM",titel="PAM corr factor")# for factor
plotPAMcorr(data,sortby = "PAM",titel="PAM corr absolute") # for absolute
}
\author{
Andreas Schönberg
}
